(Issue #14) Change components to use mbstring functions for string operations

+++ tests/integration/Form/NonpersistableStoredFormTest.php
@@ -144,6 +144,8 @@ final class NonpersistableStoredFormTest extends TestCase {
 	}
 	
 	public static function tearDownAfterClass() {
-		static::$Form->delete();
+		if(!is_null(static::$Form)) {
+			static::$Form->delete();
+		}
 	}
 }
